+++ shared/src/map/layers/tiled/Tiled2dMapLayer.cpp
@@ -137,7 +137,7 @@ void Tiled2dMapLayer::forceReload() {
     }
 }
 
-void Tiled2dMapLayer::setT(int t) {
+void Tiled2dMapLayer::setT(double t) {
     curT = t;
 
     auto mapInterface = this->mapInterface;
